SYS-2850 Makes avn-proxy transactional compatible (#125)
This commit reafactors avn-proxy pallet to be compatible with v0.9.28 or
newer versions of substrate that make all extrinsics transactional.
Up to now, if an error occurred and the extrinsic got rejected, storage
changes up to that point were persistent. Avn-proxy pallet was using
this functionality to give the fees to the relayer account, when the
inner call failed to complete.
In order to persist that behaviour, this commit makes changes on the
interface and the behaviour of the system.
Proxied extrinsics that the inner call fails to execute for any reason
will be accepted and an `InnerCallFailed` event will be emitted, with the
original dispatch error. Fee charges will be persistent in storage.

If the proxied extrinsic fails validation, the behaviour remains the
same as it was prior to this PR.

The PR also updates all test that were expecting the proxied extrinsic
to fail if the inner call fails. Instead now we look for the event and
the appropriate extrinsic info.