Sys 3135 summary pallet compilation problem (#127)
# Thank you for your Pull Request

This is to resolve the summary pallet problem when compiling on the main
runtime. This fixes the problem caused in ticket SYS-3135

Before you submit a PR, please check that:

- [x] You describe the purpose of the PR, e.g.:
  - What does it do?
  - Highlight what important points reviewers should know about;
  - Indicates if there is something left for follow-up PRs.
- [ ] Documentation updated
- [x] Business logic tested successfully
- [ ] Verify First, Write Last: In Substrate development, it is
important that you always ensure preconditions are met and return errors
at the beginning. After these checks have completed, then you may begin
the function's computation.

✄
-----------------------------------------------------------------------------