Fixes benchmarks for validators manager pallet
This commit addresses the issues with the benchmarks in the validators manager pallet.
The benchmarks have been updated to improve their setup. Now, the tests will add
additional collators on top of the existing genesis configuration.
It is required to have at least two additional collators: one collator for the
resignation action and another collator for casting a vote for the action (vote_sender)
and sending the transaction to Ethereum.

The remaining collators generated randomly are not used in any action.
Additionally, this commit refactors and simplifies the interface of the benchmark tests
to enhance their readability and maintainability.