SYS-3555 Improves error handling (#222)
This commit improves the previous work by consuming and handling errors when they occur.
On the back of that, some validators manager tests started to fail as the interaction with staking pallet can fail due to reaching the candidates bound. This has been addressed with an early check so the errors are consistent with the validators manager pallet.
Finally, some unused interfaces and tests have been removed.