Add Dockerfile for Ubuntu 22.04 (jammy) builds (#232)
This commit introduces a Dockerfile tailored for building with Ubuntu 22.04, which is compatible with Debian 12 (bookworm) based OS due to its use of OpenSSL3. The Docker image generated includes the necessary dependencies.

Additionally, the README has been updated to provide clearer instructions on selecting the appropriate Dockerfile for different build scenarios.