SYS-3930: Implement Vote Processing
This PR introduces support for processing votes upon achieving consensus.
It extends extrinsics and data structures accordingly.

Once the final vote is cast, the voting round for the Ethereum range is finalized,
and the first partition is processed. Subsequent partitions are handled using the
on_idle hook in subsequent blocks.

Additionally, this PR extends the subscription mechanism of the eth-bridge pallet.
By default, pallets ignore processed events, requiring implementation of a handler
to process them.
Benchmarking for certain actions will be addressed in a separate PR.

Jira Tasks:
- SYS-3930
- SYS-3635