feat: Adds support for additional events (#487)
Extends the client implementation to support additional events during a range processing. Additional events must be from previous blocks of the range.
Extends the runtime API to provide these and the client implementation to consume it.
Adding new events is a sudo operation performed with a  new sudo extrinsic that consolidates previous entry points.

Jira ticket:
- SYS-4453