fix/eth bridge discovered event partition order (#508)
Fixes an issue in the Ord implementation of DiscoveredEvent where event
ordering was based on the event signature instead of the transaction hash.
This caused inconsistent partitioning of the same dataset due to unstable
or incorrect ordering.

Ordering is now based on block number and transaction hash to guarantee
uniqueness and determinism.

Extends the set_admin_setting extrinsic to provide the option to reset
the vote data for ethereum events range.